Removes unused code
removes code no longer used, mostly related to the old extract files
and out-of-date data item names. Removed the functions to get a list of
currencies or currency symbols because data structure is different and
they aren’t currently being used. We can add back in once we use for
something.