fixed a regression introduced in https://github.com/Automattic/wp-cldr/commit/b3584e87a114f79176b570d8b7bcc1d82990dd77 where I tried to refactor out the case statement in `initialize_locale_bucket()` to simplify code. But after that change, data items were put into the cache based on $cldr_locale but methods like `currency_name()` tried to retrieve them based on WordPress locale. This caused us to return an empty string whenever CLDR locales differ from WordPress ones.
Bad news is that we'll need to update code in two places whenever a new data item that is locale-based is added. Good news is that caching is better since items are pre-processed before putting into cache.  Addresses #67 and partially addresses #62.

cc @jblz. Since this was breaking, and complex to debug, I went ahead and committed to master without review.